+++ src/enums/serviceAliases.ts
@@ -2,8 +2,12 @@ export default {
   camPolicy: 'camPolicies',
   ccn: 'ccns',
   ccnAttachment: 'ccnAttachments',
+  customerGateway: 'customerGateways',
+  routeTable: 'routeTables',
   securityGroup: 'securityGroups',
   subnet: 'subnets',
   vpc: 'vpcInstances',
+  vpnConnection: 'vpnConnections',
   vpnGateway: 'vpnGateways',
+  vpnGatewayRoute: 'vpnGatewayRoutes',
 }