+++ src/services/vpnConnection/index.ts
@@ -1,11 +1,11 @@
-import {Service} from '@cloudgraph/sdk'
+import { Service } from '@cloudgraph/sdk'
 import BaseService from '../base'
 import format from './format'
 import getData, { serviceName } from './data'
 import getConnections from './connections'
 import { getMutation } from '../../utils'
 
-export default class TencentSubnet extends BaseService implements Service {
+export default class TencentVpnConnection extends BaseService implements Service {
   format = format.bind(this)
 
   getData = getData.bind(this)