+++ src/types/generated.ts
@@ -68,16 +68,86 @@ export type TencentCcnAttachment = TencentBaseService & {
   state?: Maybe<Scalars['String']>;
 };
 
+export type TencentCustomerGateway = TencentBaseService & {
+  createdTime?: Maybe<Scalars['String']>;
+  ipAddress?: Maybe<Scalars['String']>;
+  name?: Maybe<Scalars['String']>;
+  tags?: Maybe<Array<Maybe<TencentRawTag>>>;
+  vpnConnections?: Maybe<Array<Maybe<TencentVpnConnection>>>;
+};
+
 export type TencentKeyValue = {
   id: Scalars['String'];
   key: Scalars['String'];
   value?: Maybe<Scalars['String']>;
 };
 
+export type TencentNetworkAcl = TencentBaseService & {
+  createdTime?: Maybe<Scalars['String']>;
+  egressEntries?: Maybe<Array<Maybe<TencentNetworkAclEntry>>>;
+  ingressEntries?: Maybe<Array<Maybe<TencentNetworkAclEntry>>>;
+  name?: Maybe<Scalars['String']>;
+  subnets?: Maybe<Array<Maybe<TencentSubnet>>>;
+  vpcId?: Maybe<Scalars['String']>;
+  vpcInstance?: Maybe<Array<Maybe<TencentVpc>>>;
+};
+
+export type TencentNetworkAclEntry = {
+  action?: Maybe<Scalars['String']>;
+  cidrBlock?: Maybe<Scalars['String']>;
+  description?: Maybe<Scalars['String']>;
+  id: Scalars['String'];
+  ipv6CidrBlock?: Maybe<Scalars['String']>;
+  modifyTime?: Maybe<Scalars['String']>;
+  port?: Maybe<Scalars['String']>;
+  protocol?: Maybe<Scalars['String']>;
+};
+
 export type TencentRawTag = {
   id: Scalars['String'];
   key?: Maybe<Scalars['String']>;
   value?: Maybe<Scalars['String']>;
+};
+
+export type TencentRouteTable = TencentBaseService & {
+  associationSet?: Maybe<Array<Maybe<TencentRouteTableAssociation>>>;
+  createdTime?: Maybe<Scalars['String']>;
+  localCidrForCcn?: Maybe<Array<Maybe<TencentRouteTableLocalCidrForCcnn>>>;
+  main?: Maybe<Scalars['Boolean']>;
+  routeSet?: Maybe<Array<Maybe<TencentRouteTableRoute>>>;
+  routeTableId?: Maybe<Scalars['String']>;
+  routeTableName?: Maybe<Scalars['String']>;
+  subnets?: Maybe<Array<Maybe<TencentSubnet>>>;
+  tags?: Maybe<Array<Maybe<TencentRawTag>>>;
+  vpcInstances?: Maybe<Array<Maybe<TencentVpc>>>;
+};
+
+export type TencentRouteTableAssociation = {
+  id: Scalars['String'];
+  routeTableId?: Maybe<Scalars['String']>;
+  subnetId?: Maybe<Scalars['String']>;
+};
+
+export type TencentRouteTableLocalCidrForCcnn = {
+  cidr?: Maybe<Scalars['String']>;
+  id: Scalars['String'];
+  publishedToVbc?: Maybe<Scalars['Boolean']>;
+};
+
+export type TencentRouteTableRoute = {
+  createdTime?: Maybe<Scalars['String']>;
+  destinationCidrBlock?: Maybe<Scalars['String']>;
+  destinationIpv6CidrBlock?: Maybe<Scalars['String']>;
+  enabled?: Maybe<Scalars['Boolean']>;
+  gatewayId?: Maybe<Scalars['String']>;
+  gatewayType?: Maybe<Scalars['String']>;
+  id: Scalars['String'];
+  publishedToVbc?: Maybe<Scalars['Boolean']>;
+  routeDescription?: Maybe<Scalars['String']>;
+  routeId?: Maybe<Scalars['Int']>;
+  routeItemId?: Maybe<Scalars['String']>;
+  routeTableId?: Maybe<Scalars['String']>;
+  routeType?: Maybe<Scalars['String']>;
 };
 
 export type TencentSecurityGroup = TencentBaseService & {
@@ -115,8 +185,10 @@ export type TencentSubnet = TencentBaseService & {
   isDefault?: Maybe<Scalars['Boolean']>;
   isRemoteVpcSnat?: Maybe<Scalars['Boolean']>;
   name?: Maybe<Scalars['String']>;
+  networkAcl?: Maybe<Array<Maybe<TencentNetworkAcl>>>;
   networkAclId?: Maybe<Scalars['String']>;
   routeTableId?: Maybe<Scalars['String']>;
+  routeTables?: Maybe<Array<Maybe<TencentRouteTable>>>;
   tags?: Maybe<Array<Maybe<TencentRawTag>>>;
   totalIpAddressCount?: Maybe<Scalars['Int']>;
   vpcInstances?: Maybe<Array<Maybe<TencentVpc>>>;
@@ -143,6 +215,8 @@ export type TencentVpc = TencentBaseService & {
   ipv6CidrBlock?: Maybe<Scalars['String']>;
   isDefault?: Maybe<Scalars['Boolean']>;
   name?: Maybe<Scalars['String']>;
+  networkAcl?: Maybe<Array<Maybe<TencentNetworkAcl>>>;
+  routeTables?: Maybe<Array<Maybe<TencentRouteTable>>>;
   subnets?: Maybe<Array<Maybe<TencentSubnet>>>;
   tags?: Maybe<Array<Maybe<TencentRawTag>>>;
   vpnGateways?: Maybe<Array<Maybe<TencentVpnGateway>>>;
@@ -154,6 +228,56 @@ export type TencentVpcAssistantCidr = {
   id: Scalars['String'];
   subnetSet?: Maybe<Array<Maybe<TencentSubnet>>>;
   vpcId?: Maybe<Scalars['String']>;
+};
+
+export type TencentVpnConnection = TencentBaseService & {
+  createdTime?: Maybe<Scalars['String']>;
+  customerGateways?: Maybe<Array<Maybe<TencentCustomerGateway>>>;
+  enableHealthCheck?: Maybe<Scalars['Boolean']>;
+  encryptProto?: Maybe<Scalars['String']>;
+  healthCheckLocalIp?: Maybe<Scalars['String']>;
+  healthCheckRemoteIp?: Maybe<Scalars['String']>;
+  healthCheckStatus?: Maybe<Scalars['String']>;
+  ikeOptionsSpecification?: Maybe<TencentVpnConnectionIkeOptionsSpecification>;
+  ipsecOptionsSpecification?: Maybe<TencentVpnConnectionIpsecOptionsSpecification>;
+  name?: Maybe<Scalars['String']>;
+  netStatus?: Maybe<Scalars['String']>;
+  preShareKey?: Maybe<Scalars['String']>;
+  routeType?: Maybe<Scalars['String']>;
+  securityPolicyDatabaseSet?: Maybe<Array<Maybe<TencentVpnConnectionSecurityPolicyDatabase>>>;
+  state?: Maybe<Scalars['String']>;
+  vpnGateways?: Maybe<Array<Maybe<TencentVpnGateway>>>;
+  vpnProto?: Maybe<Scalars['String']>;
+};
+
+export type TencentVpnConnectionIkeOptionsSpecification = {
+  dhGroupName?: Maybe<Scalars['String']>;
+  exchangeMode?: Maybe<Scalars['String']>;
+  ikeSaLifetimeSeconds?: Maybe<Scalars['Int']>;
+  ikeVersion?: Maybe<Scalars['String']>;
+  localAddress?: Maybe<Scalars['String']>;
+  localFqdnName?: Maybe<Scalars['String']>;
+  localIdentity?: Maybe<Scalars['String']>;
+  propoAuthenAlgorithm?: Maybe<Scalars['String']>;
+  propoEncryAlgorithm?: Maybe<Scalars['String']>;
+  remoteAddress?: Maybe<Scalars['String']>;
+  remoteFqdnName?: Maybe<Scalars['String']>;
+  remoteIdentity?: Maybe<Scalars['String']>;
+};
+
+export type TencentVpnConnectionIpsecOptionsSpecification = {
+  encryptAlgorithm?: Maybe<Scalars['String']>;
+  integrityAlgorith?: Maybe<Scalars['String']>;
+  ipsecSaLifetimeSeconds?: Maybe<Scalars['Int']>;
+  ipsecSaLifetimeTraffic?: Maybe<Scalars['Int']>;
+  pfsDhGroup?: Maybe<Scalars['String']>;
+};
+
+export type TencentVpnConnectionSecurityPolicyDatabase = {
+  bandwidth?: Maybe<Scalars['Int']>;
+  id: Scalars['String'];
+  localCidrBlock?: Maybe<Scalars['String']>;
+  remoteCidrBlock?: Maybe<Array<Maybe<Scalars['String']>>>;
 };
 
 export type TencentVpnGateway = TencentBaseService & {
@@ -174,7 +298,9 @@ export type TencentVpnGateway = TencentBaseService & {
   type?: Maybe<Scalars['String']>;
   version?: Maybe<Scalars['String']>;
   vpcInstances?: Maybe<Array<Maybe<TencentVpc>>>;
+  vpnConnections?: Maybe<Array<Maybe<TencentVpnConnection>>>;
   vpnGatewayQuotaSet?: Maybe<Array<Maybe<TencentVpnGatewayQuota>>>;
+  vpnGatewayRoutes?: Maybe<Array<Maybe<TencentVpnGatewayRoute>>>;
   zone?: Maybe<Scalars['String']>;
 };
 
@@ -183,4 +309,16 @@ export type TencentVpnGatewayQuota = {
   cname?: Maybe<Scalars['String']>;
   id: Scalars['String'];
   name?: Maybe<Scalars['String']>;
+};
+
+export type TencentVpnGatewayRoute = TencentBaseService & {
+  createTime?: Maybe<Scalars['String']>;
+  destinationCidrBlock?: Maybe<Scalars['String']>;
+  instanceId?: Maybe<Scalars['String']>;
+  instanceType?: Maybe<Scalars['String']>;
+  priority?: Maybe<Scalars['Int']>;
+  status?: Maybe<Scalars['String']>;
+  type?: Maybe<Scalars['String']>;
+  updateTime?: Maybe<Scalars['String']>;
+  vpnGateways?: Maybe<Array<Maybe<TencentVpnGateway>>>;
 };